backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-pptx==0.6.23
python-multipart==0.0.6
pydantic==2.5.0